user authentication update

+++ controller/todoController.js
@@ -2,7 +2,8 @@ const Todo = require("./../modles/todoModel");
 
 exports.getAllTodos = async (req, res) => {
   try {
-    const todo = await Todo.find();
+    const { userId } = req.query;
+    const todo = await Todo.find({ userId: userId });
 
     res.status(200).json({
       status: "success",
@@ -39,10 +40,13 @@ exports.getTodos = async (req, res) => {
 
 exports.createTodos = async (req, res) => {
   try {
+    const { text, userId } = req.body;
     const newTodo = await Todo.create({
-      text: req.body.text,
-      completed:false
+      text,
+      userId,
+      completed: false
     });
+
     res.status(201).json({
       status: "success",
       data: {
@@ -60,10 +64,12 @@ exports.createTodos = async (req, res) => {
 
 exports.updateTodos = async (req, res) => {
   try {
-    const todo = await Todo.findByIdAndUpdate(req.params.id, req.body, {
+    const todo = await Todo.findOneAndUpdate({_id:req.params.id, userId: req.body.userId || req.query.userId }, 
+      req.body, {
       new: true,
       runValidators: true
     });
+
     if (!todo) {
       return res.status(404).json({
         status: "fail",
@@ -88,11 +94,20 @@ exports.updateTodos = async (req, res) => {
 
 exports.deleteTodos = async (req, res) => {
   try {
-    await Todo.findByIdAndDelete(req.params.id);
+    await Todo.findOneAndDelete({ 
+      _id: req.params.id, 
+      userId: req.body.userId || req.query.userId 
+    });
+
+    if (!todo) {
+      return res.status(404).json({
+        status: "fail",
+        message: "No task found or unauthorized"
+      });
+    }
 
     res.status(204).json({
       status: "success",
-
       data: null
     });
   } catch (err) {
@@ -105,7 +120,15 @@ exports.deleteTodos = async (req, res) => {
 
 exports.deleteMany = async (req, res) => {
   try {
-    await Todo.deleteMany({});
+    const { userId } = req.query;
+
+    if (!userId) {
+      return res.status(400).json({
+        status: "fail",
+        message: "User identity required"
+      });
+    }
+    await Todo.deleteMany({userId: userId});
 
     res.status(204).json({
       status: "success",
